Restore verify option for xbutil validate on NPU.
Any change to a public API, and xbutil command line options is exactly
that, must be staged. We have to remove `verify` is multiple steps:
(1) introduce the new options while keep the old, (2) push the new
options, (3) work with clients to move them to new options, (4) remove
the old option, (5) push the removal, (6) blame the clients.

Signed-off-by: Soren Soe <2106410+stsoe@users.noreply.github.com>